“refact:decouple perf function from main code”
