+++ tools/omni/omni.h
@@ -1,5 +1,6 @@
 #include "ggml.h"
 #include "llama.h"
+#include "omni-perf.h"
 #include "tts-condition-graph.h"
 
 #include <thread>
@@ -147,12 +148,6 @@ struct projector_model {
     ggml_backend_t backend = nullptr;
     ggml_backend_buffer_type_t buf_type = nullptr;
     bool initialized = false;
-};
-
-struct OmniPerfTokenStats {
-    long long calls = 0;
-    long long tokens = 0;
-    double duration_ms = 0.0;
 };
 
 struct omni_context {
@@ -414,11 +409,7 @@ struct omni_context {
     
     // Timestamp for stream_decode start (used for WAV file naming)
     std::chrono::high_resolution_clock::time_point stream_decode_start_time;
-    std::atomic<int> perf_current_chunk_index{-1};
-    std::mutex perf_token_stats_mtx;
-    OmniPerfTokenStats perf_llm_prefill;
-    OmniPerfTokenStats perf_llm_decode;
-    OmniPerfTokenStats perf_tts_infer;
+    OmniPerfState perf;
     
     // C++ Token2Wav session for audio synthesis
     std::unique_ptr<omni::flow::Token2WavSession> token2wav_session;
@@ -489,13 +480,6 @@ struct omni_context * omni_init(struct common_params * params, int media_type, b
                                 bool duplex_mode = false,
                                 llama_model * existing_model = nullptr, llama_context * existing_ctx = nullptr,
                                 const std::string & base_output_dir = "./tools/omni/output");
-
-void omni_perf_mark(struct omni_context * ctx_omni,
-                    const char * stage,
-                    const char * event,
-                    int chunk_index = -1,
-                    double duration_ms = -1.0,
-                    const char * detail = nullptr);
 
 void omni_free(struct omni_context * ctx_omni);
